+++ lib/t_widgets_dev.dart
@@ -1,7 +1,0 @@
-library;
-
-// internal extensions (dev only)
-// export 'internal.dart';
-
-// public API
-export 't_widgets.dart';