+++ lib/theme/index.dart
@@ -1,2 +1,4 @@
 export 't_theme_services.dart';
 export 'theme_mode_listener.dart';
+export 'p_brightness_listener.dart';
+export 'p_brightness_services.dart';