+++ lib/widgets/t_loader.dart
@@ -30,7 +30,11 @@ class TLoader extends StatelessWidget {
     if (color != null) {
       return color!;
     }
-    final isDark = TWidgets.instance.getDarkMode();
+    if (isDarkMode != null) {
+      return isDarkMode! ? Colors.white : Colors.black;
+    }
+
+    final isDark = TWidgets.instance.isDarkTheme?.call() ?? false;
     return isDark ? Colors.white : Colors.black;
   }